requirements.txt
python-telegram-bot[job-queue]
selenium
logging
python-dotenv
webdriver_manager
